share_qWCP5b2WF6kIE9VVr7OPoZ/machine-cinema-newsletter/requirements.txt
pyyaml>=6.0
feedparser>=6.0.10
python-dateutil>=2.8.2
